Support for all deployment options(Yarn,Mesos,Standalone)

+++ flink-yarn/src/main/java/org/apache/flink/yarn/AbstractYarnClusterDescriptor.java
@@ -914,14 +914,13 @@ public abstract class AbstractYarnClusterDescriptor implements ClusterDescriptor
 				"");
 		}
 
-		String heapdumpDir = new File(System.getProperty("java.io.tmpdir"), appId + "-flink-heapdump").getPath();
 		final ContainerLaunchContext amContainer = setupApplicationMasterContainer(
+			appId.toString(),
 			yarnClusterEntrypoint,
 			hasLogback,
 			hasLog4j,
 			hasKrb5,
-			clusterSpecification.getMasterMemoryMB(),
-			heapdumpDir);
+			clusterSpecification.getMasterMemoryMB());
 
 		if (UserGroupInformation.isSecurityEnabled()) {
 			// set HDFS delegation tokens when security is enabled
@@ -1563,46 +1562,47 @@ public abstract class AbstractYarnClusterDescriptor implements ClusterDescriptor
 	}
 
 	protected ContainerLaunchContext setupApplicationMasterContainer(
+			String appId,
 			String yarnClusterEntrypoint,
 			boolean hasLogback,
 			boolean hasLog4j,
 			boolean hasKrb5,
-			int jobManagerMemoryMb,
-			String heapdumpDir) {
+			int jobManagerMemoryMb) {
 		// ------------------ Prepare Application Master Container  ------------------------------
 
 		String javaOpts = "";
-		// Add jvm gc options
-		String dumpFilePath = "./flink-am-heapdump.hprof";
-		String dumpFileDestPath = new File(heapdumpDir, new File(dumpFilePath).getName()).getAbsolutePath();
-		String oomScript = String.format("mkdir -p %s; mv %s %s; " +
-			"echo -e 'OutOfMemoryError! Killing current process %%p...\n" +
-			"Check gc logs and heapdump file(%s) for details.' > " +
-			ApplicationConstants.LOG_DIR_EXPANSION_VAR + "/jobmanager.err; " +
-			"kill -9 %%p",
-			heapdumpDir,
-			dumpFilePath,
-			dumpFileDestPath,
-			dumpFileDestPath);
-		String defaultGCOptions =
-			"-Xloggc:" + ApplicationConstants.LOG_DIR_EXPANSION_VAR + "/gc.log " +
-			"-XX:+PrintGCApplicationStoppedTime " +
-			"-XX:+PrintGCDetails " +
-			"-XX:+PrintGCDateStamps " +
-			"-XX:+UseGCLogFileRotation " +
-			"-XX:NumberOfGCLogFiles=10 " +
-			"-XX:GCLogFileSize=10M " +
-			"-XX:+PrintPromotionFailure " +
-			"-XX:+PrintGCCause";
-		javaOpts += defaultGCOptions;
-		if (heapdumpDir != null) {
+		// Add default gc logging options if enabled
+		boolean enableGCLogging = flinkConfiguration.getBoolean(CoreOptions.FLINK_JVM_DEFAULT_GC_LOGGING);
+		if (enableGCLogging) {
+			String defaultGCOptions =
+				"-Xloggc:" + ApplicationConstants.LOG_DIR_EXPANSION_VAR + "/gc.log " +
+					"-XX:+PrintGCApplicationStoppedTime " +
+					"-XX:+PrintGCDetails " +
+					"-XX:+PrintGCDateStamps " +
+					"-XX:+UseGCLogFileRotation " +
+					"-XX:NumberOfGCLogFiles=10 " +
+					"-XX:GCLogFileSize=10M " +
+					"-XX:+PrintPromotionFailure " +
+					"-XX:+PrintGCCause";
+			javaOpts += defaultGCOptions;
+		}
+		// Add default heap dump options if enabled
+		boolean enableHeapDump = flinkConfiguration.getBoolean(CoreOptions.FLINK_JVM_HEAPDUMP_ON_OOM);
+		String heapdumpDir = flinkConfiguration.getString(CoreOptions.FLINK_JVM_HEAPDUMP_DIRECTORY);
+		if (enableHeapDump) {
+			String dumpFileName = "flink-jm-heapdump.hprof";
+			String dumpFileDestPath = new File(heapdumpDir, appId + "-" + dumpFileName).getAbsolutePath();
+			String oomScript = String.format("echo -e 'OutOfMemoryError! Killing current process %%p...\n" +
+					"Check gc logs and heapdump file(%s) for details.' > " +
+					ApplicationConstants.LOG_DIR_EXPANSION_VAR + "/jobmanager.err; " +
+					"kill -9 %%p",
+				dumpFileDestPath);
 			javaOpts += String.format(
 				" -XX:+HeapDumpOnOutOfMemoryError " +
-				"-XX:HeapDumpPath=%s " +
-				"-XX:OnOutOfMemoryError=\"%s\"",
-				dumpFilePath,
-				oomScript
-			);
+					"-XX:HeapDumpPath=%s " +
+					"-XX:OnOutOfMemoryError=\"%s\"",
+				dumpFileDestPath,
+				oomScript);
 		}
 		// respect custom JVM options in the YAML file
 		javaOpts += " " + flinkConfiguration.getString(CoreOptions.FLINK_JVM_OPTIONS);
